… threads.

In cases where the application is set-up to have multiple routes, the route_key string may be accessed by multiple threads and cause intermittent 404 errors (e.g. one thread would enter the function, write the value and then do all of its tests based on a different value written by a different thread).

I setup two routes: `/` and `/another`. Most of the requests are served just fine, but some fail with 404:
Here's an example of a 404 captured via tcpdump for a route that does exist: 

```
10:10:37.804816 IP (tos 0x0, ttl 64, id 48100, offset 0, flags [DF], proto TCP (6), length 94)
    ip-172-31-5-244.eu-west-1.compute.internal.42704 > ip-172-31-5-65.eu-west-1.compute.internal.irdmi: Flags [P.], cksum 0x63c4 (incorrect -> 0x125b), seq 85:127, ack 317, win 227,
 options [nop,nop,TS val 324257 ecr 173515], length 42
E..^..@.@..B.......A...@..I..$.g....c......
........GET / HTTP/1.1
Host: 172.31.5.65:8000


10:10:37.804921 IP (tos 0x0, ttl 64, id 21987, offset 0, flags [DF], proto TCP (6), length 212)
    ip-172-31-5-65.eu-west-1.compute.internal.irdmi > ip-172-31-5-244.eu-west-1.compute.internal.42704: Flags [P.], cksum 0x8a7a (correct), seq 317:477, ack 127, win 210, options [n
op,nop,TS val 173515 ecr 324257], length 160
E...U.@.@......A.....@...$.g..I......z.....
........HTTP/1.1 404 Not
Server: Haywire/master
Date: Thu Dec 17 10:10:37 2015
Content-Type: text/html
Connection: Keep-Alive
Content-Length: 16

404 Not Found
```

These changes I'm proposing have a marginal positive impact on performance (results on m4.xlarge on AWS with 4 service threads, using another m4.xlarge as client), as we don't need to do a `strcpy` anymore:

### Without fix
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.31ms  159.20us  18.45ms   96.77%
    Req/Sec   761.14     67.25     2.12k    97.87%
  Latency Distribution
     50%    1.32ms
     75%    1.35ms
     90%    1.38ms
     99%    1.46ms
  384874 requests in 5.10s, 57.96MB read
Requests/sec:  75469.07
Transfer/sec:     11.37MB
```

### With fix
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 100 -t 100 -d 5s -s pipelined_get.lua --latency http://172.31.5.65:8000
Running 5s test @ http://172.31.5.65:8000
  100 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.30ms  153.82us  14.78ms   96.21%
    Req/Sec   765.15    114.88     3.32k    98.09%
  Latency Distribution
     50%    1.32ms
     75%    1.34ms
     90%    1.36ms
     99%    1.45ms
  386687 requests in 5.10s, 58.27MB read
Requests/sec:  75822.97
Transfer/sec:     11.43MB
```