@nmdguerreiro 2015-12-17
@kellabyte I'm assuming there's no reason for having a global copy of `route`. Am I missing something?